requirements.txt
# Web / scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1

# Domain enrichment
python-whois==0.9.4

# RAG — retrieval
rank-bm25==0.2.2
chromadb==0.5.3

# RAG — embeddings + LLM
voyageai==0.3.7
anthropic==0.29.0

# Output schema
pydantic==2.7.4

# UI
gradio==4.44.1
huggingface_hub==0.23.4   # gradio 4.x requires HfFolder, removed in hf_hub>=1.0

# Eval
pandas==2.2.2

# Utilities
python-dotenv==1.0.1
